requirements.txt
dash==1.4.1
dash-core-components==1.3.1
dash-html-components==1.0.1
dash-renderer==1.1.2
dash-table==4.4.1
pandas==1.0.3
Pillow==7.1.2
plotly==4.7.1
